🚧 useRateを計算するロジックとuseRateを保持する場所を大元であるChartAreaコンポーネントに移した
useRateオブジェクト達をChartとAnalysedMessageの両方で使いたかったので
